requirements.txt
pandas
openpyxl
unidecode
pyodbc
sqlalchemy
python-dotenv
ipython-sql